pyproject.toml
[project]
name = "wealth-holdings"
version = "0.1.0"
description = "负责研究和整理资料的 Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "jinja2>=3.1.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.20",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.30.0",
    "wikipedia>=1.4.0",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9.0",
]
